chore: lift out computeUpdates from stash adapter
