+++ packages/store-sync/src/exports/internal.ts
@@ -3,6 +3,7 @@ export * from "../sql";
 
 // Stash
 export * from "../stash/common";
+export * from "../stash/computeUpdates";
 export * from "../stash/createStorageAdapter";
 export * from "../stash/createSyncAdapter";
 export * from "../stash/syncToStash";
